(╯°□°）╯︵ ǝɯıʇ
So I apparently broke some minor shit the last time I fixed stuff to get the session shit handled more nicely, which apparently led to time values that should be POSIXct to be coerced to characters because I was using nested lists or arrays or whatever instead of the glorious dataframe masterrace, which meant in my sleep-deprived state that I was running and rerunning the same shit over and over trying to debug this clusterfuck of a shit forsaken wookie diarrhea and get that shit sorted out, consequently fucking a whole bunch of shit up so bad I understood this mess even less than before, but now it's probably less shitty and maybe even working, so there, have a session data update.

TL;DR: Don't go debugging when you're too tired to understand what you read, write or think.